+++ src/AIPrompt.Data/Migrations/20260717162427_AddRoadmapProjectDescription.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using AIPrompt.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
@@ -10,9 +11,11 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace AIPrompt.Data.Migrations
 {
     [DbContext(typeof(AIPromptDbContext))]
-    partial class AIPromptDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260717162427_AddRoadmapProjectDescription")]
+    partial class AddRoadmapProjectDescription
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder.HasAnnotation("ProductVersion", "8.0.11");